+++ tests/test_core.py
@@ -56,15 +56,14 @@ def test_context_log_priority_get_mock(context):
     with pytest.patch_libudev(get_prio) as get_prio:
         get_prio.return_value = mock.sentinel.log_priority
         assert context.log_priority is mock.sentinel.log_priority
-        get_prio.assert_called_with(context._context)
+        get_prio.assert_called_with(context)
 
 
 def test_context_log_priority_set_mock(context):
     set_prio = 'udev_set_log_priority'
     with pytest.patch_libudev(set_prio) as set_prio:
         context.log_priority = mock.sentinel.log_priority
-        set_prio.assert_called_with(context._context,
-                                    mock.sentinel.log_priority)
+        set_prio.assert_called_with(context, mock.sentinel.log_priority)
 
 
 def test_context_log_priority_roundtrip(context):